Full-function version routers 1.0.0
I basically completed all the functions of the routing, and some of the changes I need to be completed in the week, I also added a preliminary documentation, you can view. The documentation will be further completed, and today I am too tired to rest. All features should be available.